backend/requirements.txt
pandas
numpy
torch
scikit-learn
matplotlib
seaborn
python-dotenv
fastapi
uvicorn
groq